committed

+++ src/practice_problem_01/area_of_circle.java
@@ -7,7 +7,7 @@ public class area_of_circle
     public static void main(String[] args)
     {
         Scanner sc=new Scanner(System.in);
-        double radius = sc.nextInt();
+        double radius = sc.nextDouble();
         double ans=3.1416*radius*radius;
         System.out.println(ans);
     }
