+++ src/practice_problem_01/area_of_equilateraltriangle.java
@@ -7,7 +7,7 @@ public class area_of_equilateraltriangle
     public static void main(String[] args)
     {
         Scanner sc=new Scanner(System.in);
-        double a = sc.nextInt();
+        double a = sc.nextDouble();
         double r=Math.sqrt(3);
         double ans=a*a*r;
         ans/=4;
