+++ src/practice_problem_01/celcius_to_Fahrenheit.java
@@ -7,7 +7,7 @@ public class celcius_to_Fahrenheit
     public static void main(String[] args)
     {
         Scanner sc=new Scanner(System.in);
-        double c = sc.nextInt();
+        double c = sc.nextDouble();
         double ans=(c*9)/5;
         ans+=32;
         System.out.println(ans);
