+++ src/practice_problem_01/volume_of_a_ball.java
@@ -7,7 +7,7 @@ public class volume_of_a_ball
     public static void main(String[] args)
     {
         Scanner sc=new Scanner(System.in);
-        double radius = sc.nextInt();
+        double radius = sc.nextDouble();
         double ans=4*3.1416*radius*radius*radius;
         ans/=3;
         System.out.println(ans);
